src/python/LangChainAPI/requirements.txt
fastapi==0.103.2
uvicorn==0.23.2
azure-identity==1.14.0
azure-keyvault-secrets==4.7.0
azure-storage-blob==12.18.2
langchain==0.0.312
langchain-experimental==0.0.25
openai==0.28.1
pandas==2.1.1
tabulate==0.9.0
tiktoken==0.5.1
pymysql==1.1.0
pyodbc==4.0.39
psycopg2==2.9.9
sqlalchemy==2.0.21
numpy==1.26.0
duckduckgo-search==3.9.3
email-validator==2.0.0.post2
flask-sqlalchemy==3.1.1
chromadb==0.4.14
unstructured==0.10.22
unstructured[all-docs]==0.10.22
azure-appconfiguration-provider==1.0.0
azure-monitor-opentelemetry==1.1.0
opentelemetry-instrumentation-fastapi==0.42b0